Reworked the config XboxValues
The XboxValue class was getting quite large so I moved it to its own file.
Also added a check at the start of Config init to check if the robot has already been initialized. If so,
it should create the network table.
