+++ src/main/java/ca/team2706/frc/robot/OI.java
@@ -1,8 +1,5 @@
 package ca.team2706.frc.robot;
 
-import ca.team2706.frc.robot.commands.drivebase.ArcadeDriveWithJoystick;
-import ca.team2706.frc.robot.commands.drivebase.TankDrive;
-import ca.team2706.frc.robot.commands.drivebase.TankDriveWithJoystick;
 import ca.team2706.frc.robot.commands.drivebase.CurvatureDriveWithJoystick;
 import ca.team2706.frc.robot.config.Config;
 import ca.team2706.frc.robot.subsystems.DriveBase;
