+++ src/main/java/ca/team2706/frc/robot/commands/drivebase/TankDriveWithJoystick.java
@@ -11,25 +11,24 @@ public class TankDriveWithJoystick extends TankDrive {
     /**
      * Initializes the TankDriveWithJoystick class
      *
-     * @param joy1 The First Joystick
-     * @param axis1 The axis value of the joystick
+     * @param joy1    The First Joystick
+     * @param axis1   The axis value of the joystick
      * @param invert1 Weather or not to invert the first axis value
-     * @param joy2 The Second Joystick
-     * @param axis2 The axis value of the second joystick
+     * @param joy2    The Second Joystick
+     * @param axis2   The axis value of the second joystick
      * @param invert2 Weather or not to invert the second axis2 value
      */
     public TankDriveWithJoystick(Joystick joy1, int axis1, boolean invert1,
-                                   Joystick joy2, int axis2, boolean invert2) {
+                                 Joystick joy2, int axis2, boolean invert2) {
         super(() -> sign(joy1.getRawAxis(axis1), invert1), () -> sign(joy2.getRawAxis(axis2), invert2),
                 Config.TELEOP_SQUARE_JOYSTICK_INPUTS, Config.TELEOP_BRAKE);
     }
 
     /**
-     *
-     * @param joy The joystick to be used
-     * @param axis1 The left axis value
+     * @param joy     The joystick to be used
+     * @param axis1   The left axis value
      * @param invert1 Weather to invert the left axis value
-     * @param axis2 The right axis value
+     * @param axis2   The right axis value
      * @param invert2 Weather to invert the right axis value
      */
     public TankDriveWithJoystick(Joystick joy, int axis1, boolean invert1, int axis2, boolean invert2) {
@@ -41,6 +40,7 @@ public class TankDriveWithJoystick extends TankDrive {
         // The command should only finish when cancelled from somewhere else
         return false;
     }
+
     /**
      * Conditionally negates a number with a boolean
      *
