+++ src/main/java/ca/team2706/frc/robot/config/Config.java
@@ -13,7 +13,6 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.Objects;
 
 /**
@@ -138,6 +137,11 @@ public class Config {
         this.configTable = ntTable;
 
         Robot.setOnStateChange(this::robotStateChange);
+
+        // If the robot has already initialized itself, we should initialize constants.
+        if (Robot.isRobotInitialized()) {
+            initializeFluidConstantNetworktables();
+        }
     }
 
     /**
@@ -252,115 +256,5 @@ public class Config {
         } catch (IOException e) {
             DriverStation.reportWarning("Unable to save fluid constants to file.", false);
         }
-    }
-
-    /**
-     * Xbox controller binding information.
-     * Contains the link between the Xbox's buttons' port and the NetworkTables key used to describe the action.
-     */
-    public enum XboxValue {
-        // Axis and triggers
-        // Left on the Left Stick
-        XBOX_LEFT_STICK_X(0, "L_STICK_X", XboxInputType.Axis),
-        XBOX_LEFT_STICK_Y(1, "L_STICK_Y", XboxInputType.Axis),
-        XBOX_BACK_LEFT_TRIGGER(2, "L_TRIG", XboxInputType.Axis),
-        XBOX_BACK_RIGHT_TRIGGER(3, "R_TRIG", XboxInputType.Axis),
-        XBOX_RIGHT_STICK_X(4, "R_STICK_X", XboxInputType.Axis),
-        XBOX_RIGHT_STICK_Y(5, "R_STICK_Y", XboxInputType.Axis),
-
-        // Buttons
-        XBOX_A_BUTTON(1, "A", XboxInputType.Button),
-        XBOX_B_BUTTON(2, "B", XboxInputType.Button),
-        XBOX_X_BUTTON(3, "X", XboxInputType.Button),
-        XBOX_Y_BUTTON(4, "Y", XboxInputType.Button),
-        XBOX_LB_BUTTON(5, "LB", XboxInputType.Button),
-        XBOX_RB_BUTTON(6, "RB", XboxInputType.Button),
-        XBOX_SELECT_BUTTON(7, "SELECT", XboxInputType.Button),
-        XBOX_START_BUTTON(8, "START", XboxInputType.Button),
-        XBOX_LEFT_AXIS_BUTTON(9, "L_AXIS_BUTTON", XboxInputType.Button),
-        XBOX_RIGHT_AXIS_BUTTON(10, "R_AXIS_BUTTON", XboxInputType.Button),
-
-        // POV (The D-PAD on the XBOX Controller)
-        XBOX_POV_UP(0, "UP", XboxInputType.POV),
-        XBOX_POV_UP_RIGHT(45, "UP_RIGHT", XboxInputType.POV),
-        XBOX_POV_RIGHT(90, "RIGHT", XboxInputType.POV),
-        XBOX_POV_DOWN_RIGHT(135, "DOWN_RIGHT", XboxInputType.POV),
-        XBOX_POV_DOWN(180, "DOWN", XboxInputType.POV),
-        XBOX_POV_DOWN_LEFT(225, "DOWN_LEFT", XboxInputType.POV),
-        XBOX_POV_LEFT(270, "LEFT", XboxInputType.POV),
-        XBOX_POV_UP_LEFT(315, "UP_LEFT", XboxInputType.POV);
-
-        private final String NTString;
-        private final int port;
-        private final XboxInputType inputType;
-
-        XboxValue(int port, String NTString, XboxInputType inputType) {
-            this.NTString = NTString;
-            this.port = port;
-            this.inputType = inputType;
-        }
-
-        /**
-         * @return the nTString
-         */
-        public String getNTString() {
-            return NTString;
-        }
-
-        /**
-         * @return the port
-         */
-        public int getPort() {
-            return port;
-        }
-
-        /**
-         * Gets the input type of the input
-         *
-         * @return Either Axis, Button, or POV
-         */
-        public XboxInputType getInputType() {
-            return inputType;
-        }
-
-
-        // Create a hashmap of the networktables entry and the
-        private static final HashMap<String, XboxValue> nameMap = new HashMap<>();
-
-        static {
-            for (XboxValue value : XboxValue.values()) {
-                nameMap.put(value.getNTString(), value);
-            }
-        }
-
-        /**
-         * Gets the XboxValue constant with the given NetworkTables key.
-         *
-         * @param ntKey The NetworkTables key for the constant.
-         * @return The constant object.
-         */
-        public static XboxValue getXboxValueFromNTKey(final String ntKey) {
-            return nameMap.get(ntKey);
-        }
-    }
-
-    /**
-     * The type of input
-     */
-    public enum XboxInputType {
-        /**
-         * A raw axis or analog stick
-         */
-        Axis,
-
-        /**
-         * A button on the robot
-         */
-        Button,
-
-        /**
-         * A POV or D-pad on the robot
-         */
-        POV
     }
 }
