+++ src/test/java/ca/team2706/frc/robot/commands/drivebase/MotionMagicTest.java
@@ -16,8 +16,6 @@ import edu.wpi.first.wpilibj.command.Scheduler;
 import edu.wpi.first.wpilibj.drive.DifferentialDrive;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import mockit.*;
-import util.Util;
-
 import org.junit.Before;
 import org.junit.Test;
 import util.Util;
