+++ src/test/java/ca/team2706/frc/robot/input/FluidButtonTest.java
@@ -1,7 +1,7 @@
 package ca.team2706.frc.robot.input;
 
-import ca.team2706.frc.robot.config.Config;
 import ca.team2706.frc.robot.config.FluidConstant;
+import ca.team2706.frc.robot.config.XboxValue;
 import edu.wpi.first.wpilibj.GenericHID;
 import mockit.Expectations;
 import mockit.Injectable;
@@ -27,27 +27,27 @@ public class FluidButtonTest {
     @Tested
     public void testButtonGets() {
         new Expectations() {{
-            genericHID.getRawButton(Config.XboxValue.XBOX_A_BUTTON.getPort());
+            genericHID.getRawButton(XboxValue.XBOX_A_BUTTON.getPort());
             returns(false, true, true);
-            genericHID.getRawAxis(Config.XboxValue.XBOX_LEFT_AXIS_BUTTON.getPort());
+            genericHID.getRawAxis(XboxValue.XBOX_LEFT_AXIS_BUTTON.getPort());
             returns(0.5, -0.8, 0.9, -0.9, 0.8, 0.5);
             genericHID.getPOV(FluidButton.POV_NUMBER);
-            returns(Config.XboxValue.XBOX_POV_UP, 0, Config.XboxValue.XBOX_POV_RIGHT);
+            returns(XboxValue.XBOX_POV_UP, 0, XboxValue.XBOX_POV_RIGHT);
 
             binding.value();
             returns(
-                    Config.XboxValue.XBOX_A_BUTTON.getNTString(),
-                    Config.XboxValue.XBOX_A_BUTTON.getNTString(),
-                    Config.XboxValue.XBOX_A_BUTTON.getNTString(),
-                    Config.XboxValue.XBOX_LEFT_AXIS_BUTTON.getNTString(),
-                    Config.XboxValue.XBOX_LEFT_AXIS_BUTTON.getNTString(),
-                    Config.XboxValue.XBOX_LEFT_AXIS_BUTTON.getNTString(),
-                    Config.XboxValue.XBOX_LEFT_AXIS_BUTTON.getNTString(),
-                    Config.XboxValue.XBOX_LEFT_AXIS_BUTTON.getNTString(),
-                    Config.XboxValue.XBOX_LEFT_AXIS_BUTTON.getNTString(),
-                    Config.XboxValue.XBOX_POV_UP.getNTString(),
-                    Config.XboxValue.XBOX_POV_UP.getNTString(),
-                    Config.XboxValue.XBOX_POV_UP.getNTString());
+                    XboxValue.XBOX_A_BUTTON.getNTString(),
+                    XboxValue.XBOX_A_BUTTON.getNTString(),
+                    XboxValue.XBOX_A_BUTTON.getNTString(),
+                    XboxValue.XBOX_LEFT_AXIS_BUTTON.getNTString(),
+                    XboxValue.XBOX_LEFT_AXIS_BUTTON.getNTString(),
+                    XboxValue.XBOX_LEFT_AXIS_BUTTON.getNTString(),
+                    XboxValue.XBOX_LEFT_AXIS_BUTTON.getNTString(),
+                    XboxValue.XBOX_LEFT_AXIS_BUTTON.getNTString(),
+                    XboxValue.XBOX_LEFT_AXIS_BUTTON.getNTString(),
+                    XboxValue.XBOX_POV_UP.getNTString(),
+                    XboxValue.XBOX_POV_UP.getNTString(),
+                    XboxValue.XBOX_POV_UP.getNTString());
         }};
 
         assertFalse(fluidButton.get());
